auto-improve: Remove standalone `cai unblock` command — cycle now auto-unblocks

Refs damien-robotsix/robotsix-cai#689
